Change the package of the Record class in the Shared module
The current situation causes "java.lang.IncompatibleClassChangeError: Implementing class" due to a different class with the same name and package already being present inside the BimServer module.
