+++ Shared/src/org/bimserver/shared/SingleRecordWaitingObject.java
@@ -1,6 +1,5 @@
 package org.bimserver.shared;
 
-import org.bimserver.database.Record;
 import org.eclipse.emf.ecore.EStructuralFeature;
 
 public class SingleRecordWaitingObject extends RecordWaitingObject{
